+++ examples/apply-to/app.js
@@ -2,7 +2,7 @@
     var buttons = document.querySelectorAll('button')
 
     // Apply from stylesheet rule.
-    var sheet = jss.createStyleSheet(window.styles, true).attach()
+    var sheet = jss.createStyleSheet(window.styles).attach()
     sheet.getRule('button1').applyTo(buttons[0])
 
     // Apply from separate rule.