+++ examples/topcoat/main.js
@@ -1,5 +1,5 @@
 var jss = require('../..')
-var normalize = jss.createStyleSheet(require('./normalize')).attach()
+var normalize = jss.createStyleSheet(require('./normalize'), {named: false}).attach()
 var hogan = window.Hogan;
 
 function $(selector) {
@@ -7,8 +7,8 @@ function $(selector) {
 }
 
 (function () {
-    var buttonBar = jss.createStyleSheet(require('./topcoat-button-bar'), true).attach()
-    var button = jss.createStyleSheet(require('./topcoat-button'), true).attach()
+    var buttonBar = jss.createStyleSheet(require('./topcoat-button-bar')).attach()
+    var button = jss.createStyleSheet(require('./topcoat-button')).attach()
     var tplEl = $('.button-bar-tpl')[0]
     var buttonBarTpl = Hogan.compile(tplEl.innerHTML)